backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pypdf
pytesseract
Pillow
yt-dlp
openai-whisper
langchain
langchain-community
langchain-huggingface
faiss-cpu
google-generativeai